+++ tests/Features/Methods/ModelExtension.php
@@ -4,10 +4,30 @@ declare(strict_types=1);
 
 namespace Tests\Features\Methods;
 
+use App\User;
 use Illuminate\Database\Eloquent\Model;
 
 class ModelExtension
 {
+    /**
+     * @return iterable<\App\User>|\Illuminate\Database\Eloquent\Collection
+     */
+    public function testAll()
+    {
+        return User::all();
+    }
+
+    /**
+     * @return \App\User
+     */
+    public function testReturnThis(): User
+    {
+        $user = User::join('tickets.tickets', 'tickets.tickets.id', '=', 'tickets.sale_ticket.ticket_id')
+            ->where(['foo' => 'bar']);
+
+        return $user;
+    }
+
     public function testWhere(): Thread
     {
         return (new Thread)->where(['foo' => 'bar']);